# api/models/SwarmRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Manna AI Agent API
 * Local-first agentic AI platform. Exposes the agent reasoning loop, IDE-specific direct-LLM endpoints, file-upload helpers, and a multi-library PDF ingestion and semantic search system.  **Base URL (default):** `http://localhost:3001` 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SwarmRequest
 */
export interface SwarmRequest {
    /**
     * Natural-language description of the complex task to solve.
     * @type {string}
     * @memberof SwarmRequest
     */
    task: string;
    /**
     * When `true`, unlocks write tools for all subtask agents.
     * @type {boolean}
     * @memberof SwarmRequest
     */
    allowWrite?: boolean;
    /**
     * Force a specific model profile for all subtask agents,
     * overriding the decomposer's per-subtask suggestions.
     * 
     * @type {string}
     * @memberof SwarmRequest
     */
    profile?: SwarmRequestProfileEnum;
    /**
     * Maximum number of subtasks the decomposer may produce.
     * @type {number}
     * @memberof SwarmRequest
     */
    maxSubtasks?: number;
}


/**
 * @export
 */
export const SwarmRequestProfileEnum = {
    Fast: 'fast',
    Reasoning: 'reasoning',
    Code: 'code',
    Default: 'default'
} as const;
export type SwarmRequestProfileEnum = typeof SwarmRequestProfileEnum[keyof typeof SwarmRequestProfileEnum];


/**
 * Check if a given object implements the SwarmRequest interface.
 */
export function instanceOfSwarmRequest(value: object): value is SwarmRequest {
    if (!('task' in value) || value['task'] === undefined) return false;
    return true;
}

export function SwarmRequestFromJSON(json: any): SwarmRequest {
    return SwarmRequestFromJSONTyped(json, false);
}

export function SwarmRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SwarmRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'task': json['task'],
        'allowWrite': json['allowWrite'] == null ? undefined : json['allowWrite'],
        'profile': json['profile'] == null ? undefined : json['profile'],
        'maxSubtasks': json['maxSubtasks'] == null ? undefined : json['maxSubtasks'],
    };
}

  export function SwarmRequestToJSON(json: any): SwarmRequest {
      return SwarmRequestToJSONTyped(json, false);
  }

  export function SwarmRequestToJSONTyped(value?: SwarmRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'task': value['task'],
        'allowWrite': value['allowWrite'],
        'profile': value['profile'],
        'maxSubtasks': value['maxSubtasks'],
    };
}