# openapi.yaml
openapi: '3.1.0'
info:
    title: Manna AI Agent API
    version: '1.0.0'
    contact:
        name: Manna API
        url: https://github.com/Guebbit/manna
    description: |
        Local-first agentic AI platform. Exposes the agent reasoning loop, IDE-specific
        direct-LLM endpoints, file-upload helpers,
        and a multi-library PDF ingestion and semantic search system.

        **Base URL (default):** `http://localhost:3001`
servers:
    - url: http://localhost:3001
      description: Local development server
...
paths:
...
    /run:
        post:
            tags: [core]
            summary: Submit a task to the agent reasoning loop
            description: |
                Triggers the full agentic loop: model routing → tool selection → execution → repeat
                (up to `MAX_STEPS`, default 5). Returns the agent's final answer as a string.

                This is the right endpoint whenever no specialised endpoint covers the use case.
            operationId: postRun
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RunRequest'
                        examples:
                            basic:
                                summary: Simple read task
                                value:
                                    task: 'What npm scripts are available in this project?'
                            withProfile:
                                summary: Force reasoning profile
                                value:
                                    task: 'Explain the trade-offs of the current memory strategy'
                                    profile: reasoning
                            withWrite:
                                summary: Enable write tools
                                value:
                                    task: "Create a new file at data/hello.txt with the content 'hello'"
                                    allowWrite: true
            responses:
                '200':
                    description: Agent completed successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RunResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /run/stream:
        post:
            tags: [core]
            summary: Submit a task and receive Server-Sent Events
            description: |
                Same as `POST /run` but streams agent lifecycle events in real time as
                Server-Sent Events (SSE).

                The connection stays open until the agent completes (on `done`, `error`, or
                `max_steps` event). The original `POST /run` endpoint is completely unchanged.

                **SSE event types**:
                - `step` — emitted after each LLM response: `{ step, action, thought }`
                - `tool` — emitted after each tool execution: `{ tool, result? }` or `{ tool, error }`
                - `route` — emitted when a model profile is selected: `{ profile, model, reason }`
                - `done` — final answer ready: `{ result }`
                - `error` — agent threw: `{ error }`
                - `max_steps` — loop exhausted: `{ task, summary }`
            operationId: postRunStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RunRequest'
                        examples:
                            basic:
                                summary: Stream a simple read task
                                value:
                                    task: 'List all TypeScript files in packages/agent/'
            responses:
                '200':
                    description: SSE stream — connection kept open until agent completes
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: Newline-delimited SSE frames
                            example: |
                                event: route
                                data: {"profile":"code","model":"qwen2.5-coder:14b-instruct-q8_0","reason":"keyword_match:code"}

                                event: step
                                data: {"step":0,"action":"shell","thought":"I need to list TypeScript files..."}

                                event: tool
                                data: {"tool":"shell","result":"packages/agent/agent.ts\n..."}

                                event: done
                                data: {"result":"The TypeScript files in packages/agent/ are: ..."}

                '400':
                    $ref: '#/components/responses/BadRequest'

    # ── Swarm endpoints ───────────────────────────────────────────────────────
    /run/swarm:
        post:
            tags: [swarm]
            summary: Run a multi-agent swarm to solve a complex task
            description: |
                Decomposes the task into subtasks using an LLM, executes each subtask
                through a specialised agent (respecting dependency ordering), and
                synthesises a final answer from all subtask results.

                Use this instead of `POST /run` when the task is complex enough to
                benefit from divide-and-conquer (e.g. "build X, write tests, and document it").
            operationId: postRunSwarm
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SwarmRequest'
                        examples:
                            basic:
                                summary: Complex multi-step task
                                value:
                                    task: 'Build a React login page with form validation, write unit tests, and create API documentation'
                            withOptions:
                                summary: With profile override and subtask limit
                                value:
                                    task: 'Refactor the auth module and update the related tests'
                                    profile: code
                                    maxSubtasks: 3
                                    allowWrite: true
            responses:
                '200':
                    description: Swarm completed successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SwarmResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /run/swarm/stream:
        post:
            tags: [swarm]
            summary: Run a multi-agent swarm with SSE streaming
            description: |
                Same as `POST /run/swarm`, but streams agent lifecycle events as
                Server-Sent Events. The connection stays open until the swarm completes.

                SSE event types: `decomposed`, `subtask_start`, `subtask_done`,
                `subtask_error`, `step`, `tool`, `route`, `done`, `error`.
            operationId: postRunSwarmStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SwarmRequest'
            responses:
                '200':
                    description: SSE event stream
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    Server-Sent Events stream. Each event is typed
                                    (decomposed, subtask_start, subtask_done, step, tool, route, done, error).
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── Workflow endpoints ─────────────────────────────────────────────────────
...
components:
...
    responses:
        BadRequest:
            description: Invalid request body or parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        InternalError:
            description: Internal server error (LLM failure, Qdrant unavailable, etc.)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    type: string
                    description: Human-readable error message.
                    example: "Library 'scientific-american' not found"
...
        RunRequest:
            type: object
            required: [task]
            properties:
                task:
                    type: string
                    description: Natural-language description of what the agent should do.
                    example: 'What npm scripts are available in this project?'
                allowWrite:
                    type: boolean
                    default: false
                    description: |
                        When `true`, unlocks the `write_file`, `scaffold_project`, and `document_ingest` tools.
                        Default `false`.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile, bypassing automatic routing.
                        If omitted, the router selects a profile based on the task text.
        RunResponse:
            type: object
            properties:
                result:
                    type: string
                    description: The agent's final answer as a plain string.

        # ── Swarm schemas ────────────────────────────────────────────────────────
        SwarmRequest:
            type: object
            required: [task]
            properties:
                task:
                    type: string
                    description: Natural-language description of the complex task to solve.
                    example: 'Build a React login page, write tests, and create documentation'
                allowWrite:
                    type: boolean
                    default: false
                    description: When `true`, unlocks write tools for all subtask agents.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile for all subtask agents,
                        overriding the decomposer's per-subtask suggestions.
                maxSubtasks:
                    type: integer
                    minimum: 1
                    maximum: 10
                    default: 6
                    description: Maximum number of subtasks the decomposer may produce.
        SwarmSubtaskResult:
            type: object
            properties:
                id:
                    type: string
                    description: Subtask identifier.
                description:
                    type: string
                    description: What this subtask was asked to do.
                profile:
                    type: string
                    description: Model profile used for this subtask.
                success:
                    type: boolean
                    description: Whether the subtask completed without error.
                answer:
                    type: string
                    description: The subtask agent's final answer.
                durationMs:
                    type: number
                    description: Wall-clock execution time in milliseconds.
                error:
                    type: string
                    description: Error message (only present when `success` is `false`).
        SwarmResponse:
            type: object
            properties:
                answer:
                    type: string
                    description: The synthesised final answer.
                result:
                    type: string
                    deprecated: true
                    description: Deprecated alias for `answer` kept for backward compatibility.
                subtaskResults:
                    type: array
                    items:
                        $ref: '#/components/schemas/SwarmSubtaskResult'
                    description: Individual results for each subtask.
                decomposition:
                    type: object
                    properties:
                        reasoning:
                            type: string
                            description: Decomposer's rationale for the subtask split.
                        subtaskCount:
                            type: integer
                            description: Number of subtasks in the decomposition.
                totalDurationMs:
                    type: number
                    description: Total wall-clock duration of the entire swarm run.

        # ── Workflow schemas ─────────────────────────────────────────────────────
...